+++ http-server-netty/src/main/java/io/micronaut/http/server/netty/multipart/NettyStreamingFileUpload.java
@@ -17,13 +17,17 @@ package io.micronaut.http.server.netty.multipart;
 
 import io.micronaut.core.annotation.Internal;
 import io.micronaut.core.async.publisher.AsyncSingleResultPublisher;
+import io.micronaut.core.async.subscriber.PublisherAsBlocking;
+import io.micronaut.core.io.buffer.ReferenceCounted;
 import io.micronaut.core.naming.NameUtils;
 import io.micronaut.core.util.functional.ThrowingSupplier;
 import io.micronaut.http.MediaType;
 import io.micronaut.http.multipart.MultipartException;
 import io.micronaut.http.multipart.PartData;
 import io.micronaut.http.multipart.StreamingFileUpload;
+import io.micronaut.http.netty.PublisherAsStream;
 import io.micronaut.http.server.HttpServerConfiguration;
+import io.netty.buffer.ByteBuf;
 import io.netty.handler.codec.http.multipart.DiskFileUpload;
 import org.reactivestreams.Publisher;
 import org.reactivestreams.Subscriber;
@@ -36,6 +40,7 @@ import reactor.core.scheduler.Schedulers;
 
 import java.io.File;
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.file.Files;
 import java.util.Optional;
@@ -121,6 +126,20 @@ public final class NettyStreamingFileUpload implements StreamingFileUpload {
             fileUpload.delete();
             return true;
         });
+    }
+
+    @Override
+    public InputStream asInputStream() {
+        PublisherAsBlocking<ByteBuf> publisherAsBlocking = new PublisherAsBlocking<>() {
+            @Override
+            protected void release(ByteBuf item) {
+                if (item instanceof ReferenceCounted rc) {
+                    rc.release();
+                }
+            }
+        };
+        subject.map(pd -> ((NettyPartData) pd).getByteBuf()).subscribe(publisherAsBlocking);
+        return new PublisherAsStream(publisherAsBlocking);
     }
 
     /**
