+++ http/src/main/java/io/micronaut/http/multipart/StreamingFileUpload.java
@@ -18,6 +18,7 @@ package io.micronaut.http.multipart;
 import org.reactivestreams.Publisher;
 
 import java.io.File;
+import java.io.InputStream;
 import java.io.OutputStream;
 
 /**
@@ -79,4 +80,14 @@ public interface StreamingFileUpload extends FileUpload, Publisher<PartData> {
      */
     Publisher<Boolean> delete();
 
+    /**
+     * Create an {@link InputStream} that reads this file. The returned stream must be closed after
+     * use. The stream may block when data isn't yet available.
+     *
+     * @return An {@link InputStream} that reads this file's contents
+     * @since 4.2.0
+     */
+    default InputStream asInputStream() {
+        throw new UnsupportedOperationException("StreamingFileUpload doesn't support asInputStream");
+    }
 }
